Generate full life table from abridged life table

Convert abridged (5-year age group) life tables to full (single-year-age) life tables using specified regression parameters or lx spline

abridged_to_full(
  dt,
  id_cols,
  regression_fits,
  regression_id_cols,
  lx_spline_start_age,
  lx_spline_end_age,
  preserve_input_ax_ages = c(0, max(dt$age_start))
)

Arguments

dt
[data.table()]
Abridged life tables with variables 'age_start', 'age_end', all id_cols, 'qx', 'ax'

id_cols
[character()]
Variables that uniquely identify rows, must include 'age_start' and 'age_end'.

regression_fits
[data.table()]
Abridged to full regression parameters. Columns include all regression_id_cols, 'intercept', 'slope'.

regression_id_cols
[character()]
Variables that uniquely identify regression parameters. Must include 'age_start' and 'age_end' and be contained by id_cols.

lx_spline_start_age
[integer(1)]
Age (inclusive) to start using lx spline rather than regression fits. Use 0 to use lx spline for all ages, or Inf to use regression results for all. Corresponds to 'age_start' column.

lx_spline_end_age
[integer(1)]
Age (non-inclusive) to end spline and begin using regression fits. Corresponds to 'age_start' column.

preserve_input_ax_ages
[integer()]
Ages to preserve the input ax values for. This is typically the first age group 0-1 and the terminal age group. Corresponds to 'age_start' column.

Value

[data.table()]
Full (single-year) life tables with columns id_cols, 'age', 'qx', 'ax'.

Details

This function includes two different methods for expanding abridged life tables to full life tables. The first method is a monotonic cubic spline over lx. Because lx is always decreasing, the monotonic spline fits the general curve and pattern well, basically splining between the knots at each abridged-year increment. However, the spline can become wild and unreasonable in age groups with dramatic changes in mortality rate over age. Internally, this function calls the function stats::spline().

The second method is a regression method to predict log full qx from log abridged qx: $$log(qx) = B0 + B1 log(qx,abridged)$$ Where the parameters \(B0\) and \(B1\) are fit using high quality single-year life tables (like those from Human Mortality Database), and passed into this function. The regression must be fit separately by single-year age, but it may be also stratified by secondary identifying variables like 'sex'. These regression id variables must be present in the input abridged life tables you are trying to expand. The regression-based method has its own challenges, where jagged qx patterns may arise because we do not take consecutive abridged qx values into account simultaneously.

The function also allows for a combination of the methods to be used, with separation on age. The age cutoffs we found to work best are: regression method for ages <15 and >100 and spline method for ages between 15 and 100. This corresponds to lx_spline_start_age = 15 and lx_spline_end_age = 100.

Default ax

Values of ax are assumed to be 0.5 for all single-year ages, except those passed into preserve_input_ax_ages argument. HMD also assumes 0.5 for single-years, but we may be introducing an inconsistency between abridged and full life table mx values that are calculated from these differing ax values. Therefore, this is an area for future methods development.

Examples

data("exampleLT")
data("fullLTpars")
regression_fits <- fullLTpars[sex == "female"]
id_cols <- c("age_start", "age_end")
dt <- abridged_to_full(
  dt = exampleLT, regression_fits = regression_fits, id_cols = id_cols,
  lx_spline_start_age = 15, lx_spline_end_age = 100,
  regression_id_cols = c("age_start", "age_end")
)