Aggregate or scale qx values over age

Given a data.table with granular age interval qx values, either aggregate qx to get wider age interval qx values or scale granular qx values to be consistent with aggregate age interval qx values already present in the data.

agg_qx(
  dt,
  id_cols,
  age_mapping,
  missing_dt_severity = "stop",
  drop_present_aggs = F
)

scale_qx(dt, id_cols, missing_dt_severity = "stop")

Arguments

dt
[data.table()]
Data to be aggregated.

Must only include columns 'qx' and those specified in id_cols

Must include 'age_start' and 'age_end' columns

id_cols
[character()]
ID columns that uniquely identify each row of dt.

age_mapping
[data.table()]
Specification of intervals to aggregate to. Required columns are 'age_start' and 'age_end'. Use "Inf" as 'age_end' for terminal age group. If scaling, this mapping is inferred from the ages included.

missing_dt_severity
[character(1)]
How severe should the consequences of missing data that prevents aggregation or scaling from occurring be? Can be either 'stop', 'warning', 'message', or 'none'. If not "stop", then only the possible aggregations or scaling is done using the available data.

drop_present_aggs
[logical(1)]
Whether to drop aggregates (or overlapping intervals) that are already present in dt before aggregating. Default is 'False' and the function errors out.

Value

[data.table()]
Aggregated or scaled qx values, has id_cols and 'qx' columns for aggregate age groups if aggregating or all age groups if scaling.

Details

Aggregation: Using probability rules, we convert qx to px (conditional survival probability), multiply these px values within a specified age range together to get the conditional probability of survival through all granular age groups, and subtract from one to get the aggregated conditional probability of death (qx). agg_qx is a wrapper for demUtils::agg().

Scaling: Convert to px-space, scale up age-hierarchy so that granular px values multiply to aggregate px values, convert back to qx-space. scale_qx is a wrapper for demUtils::scale().

This function is unique to aggregating and scaling qx over age because of the dependence between age and the definition of qx. Aggregation of qx across sex, location, or other variables should be done by first aggregating ax and mx as population-weighted means (can use demUtils::agg() and demUtils::scale() directly), then recalculating qx from mx and ax (see lifetableUtils::mx_ax_to_qx()).

See also

Examples

# Example 1: aggregate qx
dt <- data.table::data.table(
  id = c(rep(1, 5), rep(2, 5)),
  qx = c(rep(0.1, 5), rep(0.2, 5)),
  age_start = rep(seq(15, 35, 5), 2),
  age_end = rep(seq(20, 40, 5), 2)
)
age_mapping <- data.table::data.table(age_start = c(15), age_end = c(40))
dt <- agg_qx(dt, id_cols = c("id", "age_start", "age_end"),
        age_mapping = age_mapping)

# Example 2: scale qx up single-year to abridged u5 hierarchy
dt <- data.table::data.table(
  sex = "male",
  qx = c(0.03, 0.015, 0.005, 0.004, 0.001, 0.03, 0.05),
  age_start = c(0, 1, 2, 3, 4, 1, 0),
  age_end = c(1, 2, 3, 4, 5, 5, 5)
)
dt <- scale_qx(dt, id_cols = c("sex", "age_start", "age_end"))