Burkina Faso example population data and ccmpp initial estimates

Example data and initial estimates for reconstructing the female population in Burkina Faso from 1960 to 2005.

burkina_faso_data

burkina_faso_initial_estimates

Format

burkina_faso_data: [data.table()] with 'population' data.

population: [data.table()] year-sex-age-specific census counts in years after the baseline year (1975, 1985, 1995, 2005). Age groups are five-year age groups from 0 to 80+.

burkina_faso_initial_estimates: list of [data.table()] of initial estimates for each ccmpp() input. Calendar year intervals are for five-year intervals between 1960 and 2005. Age groups are five-year age groups from 0 to 80+ (except 'survival' goes up to 85+). See Section: CCMPP inputs and Section: Possible [data.table()] columns for inputs for more information on each of the inputs.

An object of class list of length 5.

CCMPP inputs

srb: [data.table()] of year-specific sex ratio at birth estimates.

asfr: [data.table()] of year-age-specific average annual single-year age-specific fertility rate estimates. Must include both maternal and non-maternal age groups that are zero.

baseline: [data.table()] of year-sex-age specific baseline year population counts.

survival: [data.table()] of year-sex-age-specific survivorship ratio estimates.

net_migration: [data.table()] of year-sex-age-specific average annual net migration proportion estimates

Possible [data.table()] columns for inputs

All contain:

value_col: [numeric()] contains value of interest for each input.

If year-specific:

year_start: [integer()] start of the calendar year interval (inclusive).

year_end: [integer()] end of the calendar year interval (exclusive).

If sex-specific:

sex: [character()] either 'female' or 'male'.

If age-specific:

age_start: [integer()] start of the age group (inclusive).

age_end: [integer()] end of the age group (exclusive).

References

Wheldon, Mark C., Adrian E. Raftery, Samuel J. Clark, and Patrick Gerland. 2013. “Reconstructing Past Populations With Uncertainty From Fragmentary Data.” Journal of the American Statistical Association 108 (501): 96–110. https://doi.org/10.1080/01621459.2012.737729.

popReconstruct R Package

See also