Example life table

Example death count and population values. Also includes mx (deaths / population), and ax and qx calculated using functions from this package.

exampleLT

Format

A data.table with columns:

age_start

Integer, abridged ages 0, 1, 5, 10, ..., 95

age_end

Integer, end of age group inclusive, 1, 5, 10, ... , 95, Inf

deaths

Death counts

population

Population counts

mx

mortality rate (deaths / population)

ax

mean person-years lived by those who die in interval. Obtained using mx_to_ax function.

qx

probability of death in interval conditional on survival to xth birthday. Obtained using mx_ax_to_qx function.

Examples

exampleLT

#>     age_start age_end deaths population           mx        ax           qx
#>  1:         0       1     72      29931 2.405533e-03 0.4997995 0.0024026417
#>  2:         1       5     17     118479 1.434853e-04 1.9998087 0.0005737767
#>  3:         5      10     12     145941 8.222501e-05 2.4998287 0.0004110405
#>  4:        10      15     14     153298 9.132539e-05 2.4998097 0.0004565227
#>  5:        15      20     31     156399 1.982110e-04 2.4995871 0.0009905640
#>  6:        20      25     36     150368 2.394126e-04 2.4995012 0.0011963470
#>  7:        25      30     42     151322 2.775538e-04 2.4994218 0.0013868066
#>  8:        30      35     59     156045 3.780961e-04 2.4992123 0.0018886945
#>  9:        35      40    101     175432 5.757216e-04 2.4988006 0.0028744690
#> 10:        40      45    157     177788 8.830742e-04 2.4981603 0.0044056377
#> 11:        45      50    240     162723 1.474899e-03 2.4969273 0.0073473704
#> 12:        50      55    368     156276 2.354808e-03 2.4950942 0.0117049980
#> 13:        55      60    525     145485 3.608619e-03 2.4924821 0.0178812951
#> 14:        60      65    861     144338 5.965165e-03 2.4875728 0.0293854249
#> 15:        65      70   1019     107574 9.472549e-03 2.4802663 0.0462586307
#> 16:        70      75   1252      83885 1.492520e-02 2.4689087 0.0719094505
#> 17:        75      80   2004      74092 2.704745e-02 2.4436683 0.1264913725
#> 18:        80      85   3396      65481 5.186237e-02 2.3920743 0.2284176413
#> 19:        85      90   4972      50605 9.825116e-02 2.2961287 0.3881424621
#> 20:        90      95   3848      22853 1.683805e-01 2.1532826 0.5691100659
#> 21:        95     Inf   2245       7728 2.905021e-01 3.4423163 1.0000000000
#>     age_start age_end deaths population           mx        ax           qx