Generate abridged life table from full life table

Convert full (single-year-age) life tables to abridged (5-year age group) life tables using standard life table aggregation functions.

full_to_abridged(dt, id_cols, abridged_ages = c(0, 1, seq(5, 110, 5)))

Arguments

dt
[data.table()]
Full life table(s), variables 'age_start', 'age_end', all vars in id_cols, and at least two of 'qx', 'ax', and 'mx'. 'dx' is used but if it is not provided it is calculated within.

id_cols
[character()]
Variables that uniquely identify observations. Must include 'age_start' and 'age_end'.

abridged_ages
[integer()]
Ages to break the single-year ages into. These are the starting ages of each age interval. Default: c(0, 1, seq(5, 110, 5)).

Value

[data.table()]
Abridged life table(s) with id_cols, 'qx', and 'ax', for abridged ages specified in abridged_ages argument.

Examples

dt <- data.table::data.table(
  age_start = c(0:110),
  age_end = c(1:110, Inf),
  location = "Canada",
  qx = .2,
  ax = .5
)
id_cols = c("age_start", "age_end", "location")
dt <- full_to_abridged(dt, id_cols)