Ax iteration method

Implement iteration method to minimize the difference across initialized ax and its expected relationship with dx. This method can be used to start with naive values of ax (like age_length/2) and arrive at more informed ax values based on an approximation that assumes a relationship between ax and the age pattern of deaths (dx) in a cohort. Since dx is often calculated using the mx --> ax --> qx method in the first place, initialization and iteration is required.

iterate_ax(dt, id_cols, n_iterations = 30L, threshold = 0.01, quiet = F)

Arguments

dt
[data.table()]
Life tables with columns: all id_cols, 'age_start', 'age_end', 'dx', 'ax', 'qx', 'mx'.

id_cols
[character()]
Columns that uniquely identify each row of dt. Must include 'age_start' and 'age_end'.

n_iterations
[integer(1)]
Maximum number of iterations to run.

threshold
[numeric(1)]
How far apart should ax and the implied ax from dx be before removal from iteration. Default: 0.01.

quiet
[logical(1)]
Whether to omit messages about iteration progress. Default is F and progress messages are given.

Value

[data.table()]
Input life tables with ax, dx, qx modified. Input mx is unchanged.

Details

Limitations: this method relies on initialized ax and dx values, and cannot produce estimates for the first and last age groups. Preston pg 45.

Note: this is not the best method for ages under 5 years. See gen_u5_ax().

References

Preston Samuel H, Patrick H, Michel G. Demography: measuring and modeling population processes. MA: Blackwell Publishing. 2001.

See also

Examples

data("austria_1992_lt")
dt <- austria_1992_lt[,.(age_start, age_end, age_length, mx, ax, qx, dx, lx)]
new_dt <- iterate_ax(dt, id_cols = c("age_start", "age_end"))

#> Iteration 1

#> Number of remaining life tables: 1

#> Min. | 1st Qu. | Median | Mean | 3rd Qu. | Max. 
#>  0.25 | 0.25 | 0.25 | 0.25 | 0.25 | 0.25

#> Iteration 2

#> Number of remaining life tables: 0

#> Iterations done

if (FALSE) {
  # plot change in qx, ax, lx, dx, mx
  dt <- merge(dt, new_dt, by = c("age_start", "age_end"))
  plot(data = dt, qx.y/qx.x ~ age_start)
  plot(data = dt, ax.y/ax.x ~ age_start)
  plot(data = dt, lx.y/lx.x ~ age_start)
  plot(data = dt, dx.y/dx.x ~ age_start)
  plot(data = dt, mx.y/mx.x ~ age_start) # expect no change in mx
}